azalea_block
Struct BlockState

pub struct BlockState {
    pub id: u32,
}

A representation of a state a block can be in.

For example, a stone block only has one state but each possible stair rotation is a different state.

Fields
id: u32

The protocol ID for the block state. IDs may change every version, so you shouldnt hard-code them or store them in databases.

Implementations

impl BlockState

pub fn max_state() -> u32

Returns the highest possible state ID.

pub fn property<P: Property>(self) -> Option<P::Value>

Get a property from this block state. Will be None if the block cant have the property.

fn is_waterlogged(block_state: azalea_block::BlockState) -> bool {
    block_state.property::<azalea_block::properties::Waterlogged>().unwrap_or_default()
}

impl BlockState

pub const AIR: BlockState = _

pub fn is_valid_state(state_id: u32) -> bool

pub fn is_air(&self) -> bool

Returns true if the block is air. This only checks for normal air, not other types like cave air.